Move common toolbar items into their own class
