+++ app/src/main/java/org/mozilla/fenix/components/toolbar/ToolbarMenu.kt
@@ -37,5 +37,5 @@ interface ToolbarMenu {
     }
 
     val menuBuilder: BrowserMenuBuilder
-    val menuToolbar: BrowserMenuItemToolbar
+    val menuToolbarNavigation: BrowserMenuItemToolbar
 }
